requirements.txt
numpy==1.20.3
pandas==1.3.4
Pillow==9.3.0
PyGithub==1.57
streamlit==1.11.0
streamlit_option_menu==0.3.2
torch==1.10.1+cu113
torchvision==0.11.2+cu113
